Reorganized for now.
* Updated readme (#14)

* Just commiting. Still WIP.

* WIP.

* Still not done ... dammit.

* IDK

* A simple show using matplotlib.

* WIP.

* WIP some code commented out.

* Still WIP, deciding between doing the slicing when training or doing it separately beforehand.

* Some more fixes ...
test_prepdata() is broken though.

* Updated requirements.

* Readme WIP.

* Deep face and emotion (#15)

* Classifier log.

* Some new iml file and deepface and emotion recognition.

* Results of the deep face and emotion.

* Some deep_face_stuff.

* common_utils.py moved to ml_common package

* Classifier moved to its own folder.

* The models have been moved.

* Classifier moved accordingly.

* Image encoder decoder moved.

* Not needed ATM.

* They are all classifier related ...

* Stuff moved back as required.

* All of them moved to ml_common

* Some stuff moved.

* Image del entropy.

* More conversion to common_utils.py

* Some stuff added for Deep Face and Emotion.
