+++ .idea/MachineVision.iml
@@ -3,7 +3,7 @@
   <component name="NewModuleRootManager">
     <content url="file://$MODULE_DIR$">
       <excludeFolder url="file://$MODULE_DIR$/.venv" />
-      <excludeFolder url="file://$MODULE_DIR$/data/CIFAR" />
+      <excludeFolder url="file://$MODULE_DIR$/classifier/data/CIFAR" />
     </content>
     <orderEntry type="jdk" jdkName="Python 3.12 virtualenv at C:\mywork\python\MachineVisionExperiments\.venv" jdkType="Python SDK" />
     <orderEntry type="sourceFolder" forTests="false" />
