+++ deepfaceAndEmotion/face_results.csv
@@ -1,0 +1,44 @@
+,box,emotion,emo_confidence,x1,y1,x2,y2,f_i,image_path,size
+0,"[91, 157, 154, 237]",happy,0,91,157,154,237,0,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_281.jpg,"(1024, 432)"
+1,"[723, 125, 763, 170]",neutral,0,723,125,763,170,1,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_281.jpg,"(1024, 432)"
+2,"[438, 135, 483, 190]",happy,0,438,135,483,190,2,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_281.jpg,"(1024, 432)"
+3,"[546, 157, 583, 209]",happy,0,546,157,583,209,3,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_281.jpg,"(1024, 432)"
+4,"[273, 20, 313, 68]",fear,0,273,20,313,68,4,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_281.jpg,"(1024, 432)"
+5,"[807, 148, 851, 199]",neutral,0,807,148,851,199,5,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_281.jpg,"(1024, 432)"
+6,"[265, 121, 318, 192]",neutral,0,265,121,318,192,6,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_281.jpg,"(1024, 432)"
+7,"[162, 30, 205, 84]",neutral,0,162,30,205,84,7,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_281.jpg,"(1024, 432)"
+8,"[632, 119, 668, 159]",happy,0,632,119,668,159,8,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_281.jpg,"(1024, 432)"
+9,"[911, 160, 965, 226]",neutral,0,911,160,965,226,9,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_281.jpg,"(1024, 432)"
+10,"[450, 39, 486, 84]",happy,0,450,39,486,84,10,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_281.jpg,"(1024, 432)"
+11,"[37, 47, 82, 104]",happy,0,37,47,82,104,11,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_281.jpg,"(1024, 432)"
+12,"[380, 47, 415, 87]",happy,0,380,47,415,87,12,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_281.jpg,"(1024, 432)"
+13,"[537, 46, 571, 86]",fear,0,537,46,571,86,13,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_281.jpg,"(1024, 432)"
+14,"[367, 115, 562, 419]",neutral,0,367,115,562,419,0,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_36.jpg,"(1024, 829)"
+15,"[763, 162, 912, 372]",angry,0,763,162,912,372,1,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_36.jpg,"(1024, 829)"
+16,"[41, 46, 258, 416]",fear,0,41,46,258,416,2,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_36.jpg,"(1024, 829)"
+17,"[604, 146, 752, 355]",neutral,0,604,146,752,355,3,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_36.jpg,"(1024, 829)"
+18,"[357, 246, 418, 342]",fear,0,357,246,418,342,0,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_925.jpg,"(1024, 768)"
+19,"[474, 95, 602, 277]",happy,0,474,95,602,277,0,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_102.jpg,"(1024, 682)"
+20,"[174, 118, 299, 299]",happy,0,174,118,299,299,1,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_102.jpg,"(1024, 682)"
+21,"[696, 160, 825, 340]",happy,0,696,160,825,340,2,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_102.jpg,"(1024, 682)"
+22,"[140, 195, 165, 225]",happy,0,140,195,165,225,0,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,"(1024, 557)"
+23,"[420, 160, 448, 195]",sad,0,420,160,448,195,1,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,"(1024, 557)"
+24,"[225, 198, 251, 227]",neutral,0,225,198,251,227,2,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,"(1024, 557)"
+25,"[557, 178, 583, 210]",neutral,0,557,178,583,210,3,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,"(1024, 557)"
+26,"[868, 152, 896, 187]",happy,0,868,152,896,187,4,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,"(1024, 557)"
+27,"[780, 152, 806, 186]",happy,0,780,152,806,186,5,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,"(1024, 557)"
+28,"[388, 194, 413, 225]",happy,0,388,194,413,225,6,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,"(1024, 557)"
+29,"[310, 187, 334, 218]",happy,0,310,187,334,218,7,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,"(1024, 557)"
+30,"[520, 181, 546, 213]",happy,0,520,181,546,213,8,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,"(1024, 557)"
+31,"[616, 183, 641, 216]",happy,0,616,183,641,216,9,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,"(1024, 557)"
+32,"[492, 160, 518, 190]",happy,0,492,160,518,190,10,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,"(1024, 557)"
+33,"[342, 169, 366, 198]",happy,0,342,169,366,198,11,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,"(1024, 557)"
+34,"[273, 176, 298, 205]",fear,0,273,176,298,205,12,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,"(1024, 557)"
+35,"[202, 172, 226, 202]",sad,0,202,172,226,202,13,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,"(1024, 557)"
+36,"[683, 169, 706, 200]",happy,0,683,169,706,200,14,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,"(1024, 557)"
+37,"[470, 190, 496, 222]",neutral,0,470,190,496,222,15,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,"(1024, 557)"
+38,"[718, 141, 743, 173]",sad,0,718,141,743,173,16,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,"(1024, 557)"
+39,"[119, 217, 143, 245]",happy,0,119,217,143,245,17,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,"(1024, 557)"
+40,"[708, 178, 733, 207]",neutral,0,708,178,733,207,18,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,"(1024, 557)"
+41,"[604, 158, 626, 185]",happy,0,604,158,626,185,19,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,"(1024, 557)"
+42,"[182, 212, 205, 239]",sad,0,182,212,205,239,20,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,"(1024, 557)"
