+++ deepfaceAndEmotion/face_results_new.csv
@@ -1,0 +1,44 @@
+,image_path,h,w,f_i,x1,y1,x2,y2,emotion,emo_confidence
+0,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_281.jpg,432,1024,0,91,157,154,237,happy,0
+1,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_281.jpg,432,1024,1,723,125,763,170,neutral,0
+2,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_281.jpg,432,1024,2,438,135,483,190,happy,0
+3,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_281.jpg,432,1024,3,546,157,583,209,happy,0
+4,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_281.jpg,432,1024,4,273,20,313,68,fear,0
+5,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_281.jpg,432,1024,5,807,148,851,199,neutral,0
+6,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_281.jpg,432,1024,6,265,121,318,192,neutral,0
+7,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_281.jpg,432,1024,7,162,30,205,84,neutral,0
+8,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_281.jpg,432,1024,8,632,119,668,159,happy,0
+9,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_281.jpg,432,1024,9,911,160,965,226,neutral,0
+10,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_281.jpg,432,1024,10,450,39,486,84,happy,0
+11,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_281.jpg,432,1024,11,37,47,82,104,happy,0
+12,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_281.jpg,432,1024,12,380,47,415,87,happy,0
+13,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_281.jpg,432,1024,13,537,46,571,86,sad,0
+14,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_36.jpg,829,1024,0,367,115,562,419,neutral,0
+15,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_36.jpg,829,1024,1,763,162,912,372,angry,0
+16,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_36.jpg,829,1024,2,41,46,258,416,fear,0
+17,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_36.jpg,829,1024,3,604,146,752,355,neutral,0
+18,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\11_Meeting_Meeting_11_Meeting_Meeting_11_925.jpg,768,1024,0,357,246,418,342,fear,0
+19,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_102.jpg,682,1024,0,474,95,602,277,happy,0
+20,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_102.jpg,682,1024,1,174,118,299,299,happy,0
+21,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_102.jpg,682,1024,2,696,160,825,340,happy,0
+22,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,557,1024,0,140,195,165,225,neutral,0
+23,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,557,1024,1,420,160,448,195,sad,0
+24,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,557,1024,2,225,198,251,227,neutral,0
+25,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,557,1024,3,557,178,583,210,neutral,0
+26,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,557,1024,4,868,152,896,187,happy,0
+27,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,557,1024,5,780,152,806,186,happy,0
+28,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,557,1024,6,388,194,413,225,happy,0
+29,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,557,1024,7,310,187,334,218,happy,0
+30,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,557,1024,8,520,181,546,213,happy,0
+31,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,557,1024,9,616,183,641,216,happy,0
+32,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,557,1024,10,492,160,518,190,happy,0
+33,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,557,1024,11,342,169,366,198,happy,0
+34,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,557,1024,12,273,176,298,205,fear,0
+35,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,557,1024,13,202,172,226,202,sad,0
+36,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,557,1024,14,683,169,706,200,happy,0
+37,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,557,1024,15,470,190,496,222,neutral,0
+38,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,557,1024,16,718,141,743,173,sad,0
+39,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,557,1024,17,119,217,143,245,happy,0
+40,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,557,1024,18,708,178,733,207,neutral,0
+41,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,557,1024,19,604,158,626,185,happy,0
+42,C:\mywork\ml\deepfaceAndEmotion\MultiFaceCheck\TestFaces\12_Group_Group_12_Group_Group_12_67.jpg,557,1024,20,182,212,205,239,sad,0
