+++ image_encoder_decoder/src/image_encoder_decoder/models/skeleton_model.py
@@ -1,0 +1,11 @@
+from torch import nn
+
+
+class Animal_Text_Encoder(nn.Module):
+	def __init__(self):
+		super().__init__()
+
+
+
+
+
