+++ requirements.txt
@@ -5,3 +5,4 @@ matplotlib~=3.10.1
 numpy~=2.1.2
 torchvision~=0.21.0+cu124
 pandas~=2.2.3
+datasets~=3.3.2
