--- src/models/skeleton_model.py
@@ -1,11 +1,0 @@
-from torch import nn
-
-
-class Animal_Text_Encoder(nn.Module):
-	def __init__(self):
-		super().__init__()
-
-
-
-
-
