--- src/wip/animal_data_preparation.py
@@ -1,4 +1,0 @@
-
-
-
-
